Move givenThreads() to __tests__/__utils__
