+++ __tests__/__utils__/handlers.ts
@@ -132,6 +132,32 @@ export function given<M extends DatabaseLayer.MessageType>(type: M) {
   }
 }
 
+export function givenThreads({
+  uuid,
+  threads,
+}: {
+  uuid: Model<'AbstractUuid'>
+  threads: Model<'Comment'>[][]
+}) {
+  const firstCommentIds = threads.map((thread) => thread[0].id)
+  const firstComments = threads.map((thread) => {
+    const childrenIds = thread.slice(1).map(R.prop('id'))
+
+    return { ...thread[0], parentId: uuid.id, childrenIds }
+  })
+  const otherComments = threads.flatMap((thread) =>
+    thread.slice(1).map((comment) => {
+      return { ...comment, parentId: thread[0].id, childrenIds: [] }
+    })
+  )
+
+  given('ThreadsQuery')
+    .withPayload({ id: uuid.id })
+    .returns({ firstCommentIds })
+
+  given('UuidQuery').for(uuid, firstComments, otherComments)
+}
+
 export function createNotificationEventHandler(
   notificationEvent: Model<'AbstractNotificationEvent'>
 ) {
